requirements.txt
groq
pycoingecko
python-dotenv
tabulate
yfinance
